refactor: update server to use modular architecture
